Move table 'structure' schemas out to main folder, since locations got own folder outside tables.
