+++ src/ogd/common/schemas/tables/TableSchema.py
@@ -3,10 +3,10 @@ import logging
 from typing import Any, Dict, List, Optional, TypeAlias
 ## import local files
 from ogd.common.schemas.Schema import Schema
-from ogd.common.schemas.tables.structures.ColumnSchema import ColumnSchema
-from ogd.common.schemas.tables.structures.TableStructureSchema import TableStructureSchema
-from ogd.common.schemas.tables.structures.EventTableStructureSchema import EventTableStructureSchema
-from ogd.common.schemas.tables.structures.FeatureTableStructureSchema import FeatureTableStructureSchema
+from ogd.common.schemas.tables.ColumnSchema import ColumnSchema
+from ogd.common.schemas.tables.TableStructureSchema import TableStructureSchema
+from ogd.common.schemas.tables.EventTableStructureSchema import EventTableStructureSchema
+from ogd.common.schemas.tables.FeatureTableStructureSchema import FeatureTableStructureSchema
 from ogd.common.schemas.locations.TableLocationSchema import URLLocationSchema
 from ogd.common.utils.Logger import Logger
 from ogd.common.utils.typing import Map
