--- src/ogd/common/schemas/tables/structures/__init__.py
@@ -1,9 +1,0 @@
-__all__ = [
-    "TableStructureSchema",
-    "EventTableStructureSchema",
-    "FeatureTableStructureSchema"
-]
-
-from . import TableStructureSchema
-from . import EventTableStructureSchema
-from . import FeatureTableStructureSchema
